requirements.txt
# django
Django==4.0.3
django-extensions==3.1.5


# djangorestframework
djangorestframework==3.13.1
django-filter==21.1


# Authentication and authorization
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-dry-rest-permissions==1.2.0  # https://github.com/FJNR-inc/dry-rest-permissions


# tests, coverage
djet==0.3.0
model-bakery==1.5.0
coverage==6.3.2


# dependency and safety checker
safety==1.10.3
insecure-package==0.1.0
graphviz==0.19.2
pipdeptree==2.2.1


# pre-commit, linter, formatter
pre_commit==2.18.1
flake8==4.0.1
flake8-bugbear==22.3.23
isort==5.10.1
autoflake==1.4
black==22.3.0
